revert back to Integer type for timeSinceLastCommit
RDF persist does not work with long values for some reason

+++ smartclide-monitoring/src/main/java/eu/smartclide/contexthandling/dle/model/CommitMessage.java
@@ -38,7 +38,7 @@ public class CommitMessage {
     String user;
     String branch;
     @SerializedName("time_since_last_commit")
-    Long timeSinceLastCommit;
+    Integer timeSinceLastCommit;
     @SerializedName("number_of_files_modified")
     Integer numberOfFilesModified;
 }
